:sparkles: add method to download more than 1 show, and seperate method for usability
