+++ Gemfile
@@ -2,11 +2,11 @@ source 'https://rubygems.org'
 
 gem 'byebug'
 gem 'httparty'
-gem 'nokogiri'
 gem 'json'
-gem "tty-prompt"
-gem 'tty-progressbar'
-gem 'pastel'
-gem 'tty-spinner'
 gem 'nikkou'
+gem 'nokogiri'
+gem 'pastel'
 gem 'tty-font'
+gem 'tty-progressbar'
+gem 'tty-prompt'
+gem 'tty-spinner'
