--- eztv-scraped/Holmes The Next Generation.json
@@ -1,43 +1,0 @@
-[
-  {
-    "title": "Holmes: The Next Generation",
-    "description": "Contractor Mike Holmes joins forces with his kids, Sherry and Mike Junior, to rescue homeowners from shady contractors, DIY projects gone wrong and dysfunctional homes.",
-    "poster": "/ezimg/thumbs/holmes-the-next-generation-3782.jpg",
-    "imdb_rating": "9.30"
-  },
-  {
-    "episode_name": "Holmes-Next Generation S01E06 Nightmares and Dreams 720p WEB x264-KOMPOST [eztv]",
-    "magnet": "magnet:?xt=urn:btih:d856d9a90161c4a03f8dce750985ac44bdcd9dec&dn=Holmes-Next.Generation.S01E06.Nightmares.and.Dreams.720p.WEB.x264-KOMPOST%5Beztv%5D.mkv%5Beztv%5D&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A80&tr=udp%3A%2F%2Fglotorrents.pw%3A6969%2Fannounce&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337%2Fannounce&tr=udp%3A%2F%2Fexodus.desync.com%3A6969",
-    "size": "914.44 MB",
-    "released": "1 year",
-    "seeds": "-"
-  },
-  {
-    "episode_name": "Holmes-Next Generation S01E05 Old is New Again 720p WEB x264-KOMPOST [eztv]",
-    "magnet": "magnet:?xt=urn:btih:040a89014586611e2c199d0f5406c10a80081c8d&dn=Holmes-Next.Generation.S01E05.Old.is.New.Again.720p.WEB.x264-KOMPOST%5Beztv%5D.mkv%5Beztv%5D&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A80&tr=udp%3A%2F%2Fglotorrents.pw%3A6969%2Fannounce&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337%2Fannounce&tr=udp%3A%2F%2Fexodus.desync.com%3A6969",
-    "size": "916.05 MB",
-    "released": "1 year",
-    "seeds": "-"
-  },
-  {
-    "episode_name": "Holmes-Next Generation S01E04 And Baby Makes Three 720p WEB x264-KOMPOST [eztv]",
-    "magnet": "magnet:?xt=urn:btih:7cba91266184cfddb04178245fca74a4422c1c5b&dn=Holmes-Next.Generation.S01E04.And.Baby.Makes.Three.720p.WEB.x264-KOMPOST%5Beztv%5D.mkv%5Beztv%5D&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A80&tr=udp%3A%2F%2Fglotorrents.pw%3A6969%2Fannounce&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337%2Fannounce&tr=udp%3A%2F%2Fexodus.desync.com%3A6969",
-    "size": "914.15 MB",
-    "released": "1 year",
-    "seeds": "-"
-  },
-  {
-    "episode_name": "Holmes-Next Generation S01E03 Far From Home 720p WEB x264-KOMPOST [eztv]",
-    "magnet": "magnet:?xt=urn:btih:90a71deddab8473d527cfbf24455d31afafab15d&dn=Holmes-Next.Generation.S01E03.Far.From.Home.720p.WEB.x264-KOMPOST%5Beztv%5D.mkv%5Beztv%5D&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A80&tr=udp%3A%2F%2Fglotorrents.pw%3A6969%2Fannounce&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337%2Fannounce&tr=udp%3A%2F%2Fexodus.desync.com%3A6969",
-    "size": "914.98 MB",
-    "released": "1 year",
-    "seeds": "-"
-  },
-  {
-    "episode_name": "Holmes-Next Generation S01E02 Hiding in Plain Sight 720p WEB x264-KOMPOST [eztv]",
-    "magnet": "magnet:?xt=urn:btih:d1b82e61ac5616f0b036e0131c635a6918a338e5&dn=Holmes-Next.Generation.S01E02.Hiding.in.Plain.Sight.720p.WEB.x264-KOMPOST%5Beztv%5D.mkv%5Beztv%5D&tr=udp%3A%2F%2Ftracker.coppersurfer.tk%3A80&tr=udp%3A%2F%2Fglotorrents.pw%3A6969%2Fannounce&tr=udp%3A%2F%2Ftracker.opentrackr.org%3A1337%2Fannounce&tr=udp%3A%2F%2Fexodus.desync.com%3A6969",
-    "size": "915.64 MB",
-    "released": "1 year",
-    "seeds": "-"
-  }
-]
