--- lib/extract.rb
@@ -1,3 +1,0 @@
-class Extract
-    
-end
